backend/auth/requirements.txt
psycopg2-binary==2.9.7
bcrypt==4.2.1
PyJWT==2.8.0